requirements.txt
Django==4.2
djangorestframework
psycopg2-binary
gunicorn
whitenoise
black
flake8
isort

